+++ src/pydvl/valuation/utility/torchutility.py
